+++ src/boundary_functions.jl
@@ -16,16 +16,16 @@ abstract type AbstractBoundaryFunction <: MOI.AbstractScalarFunction end
 ## Initial & Final
 
 """
-    Initial{DF}(dyn_fun::DF) where {DF<:AbstractDynamicFunction}
+    Initial{DF}(evaluand::DF) where {DF<:AbstractDynamicFunction}
 
 Represents the evaluation of an [`AbstractDynamicFunction`](@ref) at the initial 
 point of its phase.
 
 It is a sub-type of [`AbstractBoundaryFunction`](@ref). The dynamic function is stored
-in the `dyn_fun` field.
+in the `evaluand` field.
 """
 struct Initial{DF<:AbstractDynamicFunction} <: AbstractBoundaryFunction
-    dyn_fun::DF
+    evaluand::DF
 end
 
 function MOI.Utilities._to_string(
@@ -35,22 +35,22 @@ function MOI.Utilities._to_string(
 )
     return string(
         "Initial(",
-        MOI.Utilities._to_string(options, model, initial.dyn_fun),
+        MOI.Utilities._to_string(options, model, initial.evaluand),
         ")",
     )
 end
 
 """
-    Final{DF}(dyn_fun::DF) where {DF<:AbstractDynamicFunction}
+    Final{DF}(evaluand::DF) where {DF<:AbstractDynamicFunction}
 
 Represents the evaluation of an [`AbstractDynamicFunction`](@ref) at the final 
 point of its phase.
 
 It is a sub-type of [`AbstractBoundaryFunction`](@ref). The dynamic function is stored
-in the `dyn_fun` field.
+in the `evaluand` field.
 """
 struct Final{DF<:AbstractDynamicFunction} <: AbstractBoundaryFunction
-    dyn_fun::DF
+    evaluand::DF
 end
 
 function MOI.Utilities._to_string(
@@ -60,7 +60,7 @@ function MOI.Utilities._to_string(
 )
     return string(
         "Final(",
-        MOI.Utilities._to_string(options, model, final.dyn_fun),
+        MOI.Utilities._to_string(options, model, final.evaluand),
         ")",
     )
 end
@@ -69,18 +69,18 @@ end
 
 """
     Linkage{DF}(
-        final::Final{DF},
-        initial::Initial{DF},
+        final_evaluand::DF,
+        initial_evaluand::DF,
     ) where {DF<:AbstractDynamicFunction}
 
 Represents the expression ``f_f(y(t^f), t^f) - f_0(y(t^0), t^0)``.
 
 It is a sub-type of [`AbstractBoundaryFunction`](@ref). The final function is stored in the
-`final` field and the initial function is stored in the `initial` field.
+`final_evaluand` field and the initial function is stored in the `initial_evaluand` field.
 """
 struct Linkage{DF<:AbstractDynamicFunction} <: AbstractBoundaryFunction
-    final::Final{DF}
-    initial::Initial{DF}
+    final_evaluand::DF
+    initial_evaluand::DF
 end
 
 function MOI.Utilities._to_string(
@@ -89,9 +89,11 @@ function MOI.Utilities._to_string(
     linkage::Linkage,
 )
     return string(
-        MOI.Utilities._to_string(options, model, linkage.final),
-        " - ",
+        "Final(",
+        MOI.Utilities._to_string(options, model, linkage.final_evaluand),
+        ") - Initial(",
         MOI.Utilities._to_string(options, model, linkage.initial),
+        ")",
     )
 end
 
@@ -149,14 +151,14 @@ end
 ## Integrals
 
 """
-    Integral{DF}(dyn_fun::AbstractDynamicFunction) where {DF<:AbstractDynamicFunction}
+    Integral{DF}(integrand::DF) where {DF<:AbstractDynamicFunction}
 
 Represents the integral ``\\int_{t_i^o}^{t_i^f} f_d(\\dot{y}(t_i), y(t_i), t_i, x) \\mathrm{d}t_i``.
 
-It is a sub-type of [`AbstractBoundaryFunction`](@ref). The integrand is stored in the `dyn_fun` field.
+It is a sub-type of [`AbstractBoundaryFunction`](@ref). The integrand is stored in the `integrand` field.
 """
 struct Integral{DF<:AbstractDynamicFunction} <: AbstractBoundaryFunction
-    dyn_fun::DF
+    integrand::DF
 end
 
 function MOI.Utilities._to_string(
@@ -166,34 +168,75 @@ function MOI.Utilities._to_string(
 )
     return string(
         "∫(",
-        MOI.Utilities._to_string(options, model, integral.dyn_fun),
+        MOI.Utilities._to_string(options, model, integral.integrand),
         ")d(",
         MOI.Utilities._to_string(options, model, phase_index(integral)),
         ")",
     )
 end
 
-phase_index(integral::Integral) = phase_index(integral.dyn_fun)
+phase_index(integral::Integral) = phase_index(integral.integrand)
 
 """
-    Bolza{BF,DF}(
+    MultiPhaseIntegral{DF}(
+        integrands::Vector{DF},
+    ) where {DF<:AbstractDynamicFunction}
+
+Represents the sum of integrals
+``\\sum_i \\big[ \\int_{t_i^o}^{t_i^f} f_d(\\dot{y}(t_i), y(t_i), t_i, x) \\mathrm{d}t_i \\big]``.
+
+It is a sub-type of [`AbstractBoundaryFunction`](@ref). The vector of integrands is
+stored in the `integrands` field.
+"""
+struct MultiPhaseIntegral{DF<:AbstractDynamicFunction} <: AbstractBoundaryFunction
+    integrands::Vector{DF}
+end 
+
+function MOI.Utilities._to_string(
+    options::MOI.Utilities._PrintOptions,
+    model::MOI.ModelLike,
+    multi_phase_integral::MultiPhaseIntegral,
+)
+    integrands = multi_phase_integral.integrands
+
+    s = string(
+        "∫(",
+        MOI.Utilities._to_string(options, model, integrands[1]),
+        ")d(",
+        MOI.Utilities._to_string(options, model, phase_index(integrands[1])),
+        ")",
+    )
+
+    for i in 2:length(integrands)
+        s *= string(
+            " + ∫(",
+            MOI.Utilities._to_string(options, model, integrands[i]),
+            ")d(",
+            MOI.Utilities._to_string(options, model, phase_index(integrands[i])),
+            ")",
+        )
+    end
+    return s
+end
+
+"""
+    Bolza{BF,IF}(
         bou_fun::BF,
-        integral::Integral{DF},
-    ) where {BF<:AbstractBoundaryFunction,DF<:AbstractDynamicFunction}
+        integral::IF,
+    ) where {BF<:AbstractBoundaryFunction,IF<:Union{Integral,MultiPhaseIntegral}}
 
 ```math
 f_b(y_0, y_f, t_0, t_f, x) + \\int_{t_i^o}^{t_i^f} f_d(\\dot{y}(t_i), y(t_i), t_i, x) \\mathrm{d}t_i
 ```
-Represents the sum of an [`AbstractBoundaryFunction`](@ref) with the integral of
-an [`AbstractDynamicFunction`](@ref).
+Represents the sum of an [`AbstractBoundaryFunction`](@ref) with either an [`Integral`](@ref) or a
+[`MultiPhaseIntegral`](@ref).
 
 It is a sub-type of [`AbstractBoundaryFunction`](@ref). The boundary function is stored in the `bou_fun`
 field and the integral is stored in the `integral` field.
 """
-struct Bolza{BF<:AbstractBoundaryFunction,DF<:AbstractDynamicFunction} <:
-       AbstractBoundaryFunction
+struct Bolza{BF<:AbstractBoundaryFunction,IF<:Union{Integral,MultiPhaseIntegral}} <: AbstractBoundaryFunction
     bou_fun::BF
-    integral::Integral{DF}
+    integral::IF
 end
 
 function MOI.Utilities._to_string(