+++ test/phases.jl
@@ -10,12 +10,12 @@
     @test DOI.supports_phase(model) == false
     @test_throws DOI.AddPhaseNotAllowed DOI.add_phase(model)
     @test MOI.is_valid(model, t_1) == false
-    @test_throws MOI.UnsupportedAttribute MOI.set(model, MOI.VariableName(), t_1, "Time")
+    @test_throws MethodError MOI.set(model, MOI.VariableName(), t_1, "Time")
 
     struct PhaseModel <: MOI.ModelLike end
     MOI.supports(::PhaseModel, ::MOI.VariableName, ::Type{DOI.PhaseIndex}) = true
     model = PhaseModel()
 
-    @test_throws MOI.SetAttributeNotAllowed MOI.set(model, MOI.VariableName(), t_1, "Time")
+    @test_throws MethodError MOI.set(model, MOI.VariableName(), t_1, "Time")
     
 end