+++ docs/src/changelog.md
@@ -7,6 +7,12 @@ CurrentModule = DynOptInterface
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v0.2.0 (August 31, 2024)
+
+- Added phase attributes
+- Changed `DynamicVariableDerivative` to `Derivative`, also changing `ExplicitDifferentialFunction`
+- Added `MultiPhaseIntegral`
+
 ## v0.1.0 (July 26, 2024)
 
 - Initial release